docs: Fix links to other component props
Previously if you were on another version of the docs and you clicked on a link to take you to another component's props, it would always take you to the latest version released.

This commit fixes this so that links clicked stay on the current version you're browsing.

Also fixes lint-staged prettier task to add the files to git.